+++ lib/src/int_extension.dart
@@ -1,4 +1,7 @@
 extension IntExtension on int {
+  Duration asDays() => Duration(days: this);
+  Duration asHours() => Duration(hours: this);
+  Duration asMinutes() => Duration(minutes: this);
   Duration asSeconds() => Duration(seconds: this);
   Duration asMilliseconds() => Duration(milliseconds: this);
 }