+++ lib/src/list_extension.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 extension ListExtension<T> on List<T> {
   void addBetween(value) {
     for (int i = length - 1; i > 0; i--) {
@@ -5,11 +7,5 @@ extension ListExtension<T> on List<T> {
     }
   }
 
-  T? get nullOrFirst => isBlank ? null : first;
-  T? get nullOrLast => isBlank ? null : last;
-}
-
-extension NullListExtension<T> on List<T>? {
-  bool get isBlank => this == null || this!.isEmpty;
-  bool get isNotBlank => !isBlank;
+  String toJSON() => jsonEncode(this);
 }