+++ CHANGELOG.md
@@ -35,3 +35,11 @@
 
 - Fix scrollable widget extension
 - Fix add months extension
+
+## [0.0.14] - 08/08/2021
+
+- Change default border color
+
+## [0.0.15] - 08/08/2021
+
+- Extensions for leap year and days in month